requirements.txt
blinker==1.9.0
Bootstrap-Flask==2.5.0
click==8.2.1
colorama==0.4.6
dominate==2.9.1
Flask==3.1.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
visitor==0.1.3
psycopg2==2.9.11
Werkzeug==3.1.3
WTForms==3.2.1
